+++ Modules/Microsoft365DSC/DSCResources/MSFT_SCAuditConfigurationPolicy/MSFT_SCAuditConfigurationPolicy.psm1
@@ -20,7 +20,6 @@ function Get-TargetResource
     )
 
     Write-Verbose -Message "Getting configuration of SCAuditConfigurationPolicy for Workload {$Workload}"
-
     Write-Verbose -Message "Connecting to Security and Compliance Center"
 
     if ($Global:CurrentModeIsExport)